+++ android/src/test/java/com/instabug/flutter/InstabugApiTest.java
@@ -59,6 +59,7 @@ import java.util.Map;
 import java.util.concurrent.Callable;
 
 import io.flutter.plugin.common.BinaryMessenger;
+import org.mockito.verification.VerificationMode;
 
 public class InstabugApiTest {
     private final Callable<Bitmap> screenshotProvider = () -> mock(Bitmap.class);
@@ -155,6 +156,20 @@ public class InstabugApiTest {
         api.setEnabled(isEnabled);
 
         mInstabug.verify(Instabug::disable);
+    }
+
+    @Test
+    public void testIsEnabled() {
+        api.isEnabled();
+
+        mInstabug.verify(Instabug::isEnabled);
+    }
+
+    @Test
+    public void testIsBuilt() {
+        api.isBuilt();
+
+        mInstabug.verify(Instabug::isBuilt);
     }
 
     @Test