+++ android/src/test/java/com/instabug/flutter/util/GlobalMocks.java
@@ -8,12 +8,19 @@ import android.graphics.Bitmap;
 import android.net.Uri;
 import android.util.Log;
 
+import com.instabug.crash.models.IBGNonFatalException;
+
 import org.json.JSONObject;
 import org.mockito.MockedStatic;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 
 import java.lang.reflect.Method;
+import java.util.Map;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.mockStatic;
 
 public class GlobalMocks {
     public static MockedStatic<ThreadManager> threadManager;
@@ -75,8 +82,29 @@ public class GlobalMocks {
                         JSONObject.class, boolean.class))
                 .thenReturn(mCrashReportException);
 
+        Method mCrashReportNonFatalException = MockReflected.class.getDeclaredMethod("crashReportException", JSONObject.class, boolean.class,
+                Map.class, JSONObject.class, IBGNonFatalException.Level.class);
+        mCrashReportNonFatalException.setAccessible(true);
+        reflection
+                .when(() -> Reflection.getMethod(Class.forName("com.instabug.crash.CrashReporting"), "reportException",
+                        JSONObject.class, boolean.class,
+                        Map.class, JSONObject.class, IBGNonFatalException.Level.class))
+                .thenReturn(mCrashReportNonFatalException);
+
         uri = mockStatic(Uri.class);
         uri.when(() -> Uri.fromFile(any())).thenReturn(mock(Uri.class));
+
+        Method mStartUiTraceCP = MockReflected.class.getDeclaredMethod("startUiTraceCP", String.class, Long.class, Long.class);
+        mStartUiTraceCP.setAccessible(true);
+        reflection.when(() -> Reflection.getMethod(Class.forName("com.instabug.apm.APM"), "startUiTraceCP", String.class, Long.class, Long.class)).thenReturn(mStartUiTraceCP);
+
+        Method mReportScreenLoadingCP = MockReflected.class.getDeclaredMethod("reportScreenLoadingCP", Long.class, Long.class, Long.class);
+        mReportScreenLoadingCP.setAccessible(true);
+        reflection.when(() -> Reflection.getMethod(Class.forName("com.instabug.apm.APM"), "reportScreenLoadingCP", Long.class, Long.class, Long.class)).thenReturn(mReportScreenLoadingCP);
+
+        Method mEndScreenLoadingCP = MockReflected.class.getDeclaredMethod("endScreenLoadingCP", Long.class, Long.class);
+        mEndScreenLoadingCP.setAccessible(true);
+        reflection.when(() -> Reflection.getMethod(Class.forName("com.instabug.apm.APM"), "endScreenLoadingCP", Long.class, Long.class)).thenReturn(mEndScreenLoadingCP);
     }
 
     public static void close() {