+++ android/src/test/java/com/instabug/flutter/util/MockReflected.java
@@ -1,10 +1,13 @@
 package com.instabug.flutter.util;
 
 import android.graphics.Bitmap;
-
 import androidx.annotation.Nullable;
 
+import com.instabug.crash.models.IBGNonFatalException;
+
 import org.json.JSONObject;
+
+import java.util.Map;
 
 /**
  * Includes fake implementations of methods called by reflection.
@@ -36,4 +39,12 @@ public class MockReflected {
      * CrashReporting.reportException
      */
     public static void crashReportException(JSONObject exception, boolean isHandled) {}
+    public static void crashReportException(JSONObject exception, boolean isHandled, Map<String,String> userAttributes, JSONObject fingerPrint, IBGNonFatalException.Level level) {}
+
+
+    public static void startUiTraceCP(String screenName, Long microTimeStamp, Long traceId) {}
+
+    public static void reportScreenLoadingCP(Long startTimeStampMicro, Long durationMicro, Long uiTraceId) {}
+
+    public static void endScreenLoadingCP(Long timeStampMicro, Long uiTraceId) {}
 }