+++ example/lib/src/components/network_content.dart
@@ -1,0 +1,47 @@
+part of '../../main.dart';
+
+class NetworkContent extends StatefulWidget {
+  const NetworkContent({Key? key}) : super(key: key);
+  final String defaultRequestUrl =
+      'https://jsonplaceholder.typicode.com/posts/1';
+
+  @override
+  State<NetworkContent> createState() => _NetworkContentState();
+}
+
+class _NetworkContentState extends State<NetworkContent> {
+  final endpointUrlController = TextEditingController();
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        InstabugClipboardInput(
+          label: 'Endpoint Url',
+          controller: endpointUrlController,
+        ),
+        InstabugButton(
+          text: 'Send Request To Url',
+          onPressed: () => _sendRequestToUrl(endpointUrlController.text),
+        ),
+      ],
+    );
+  }
+
+  void _sendRequestToUrl(String text) async {
+    try {
+      String url = text.trim().isEmpty ? widget.defaultRequestUrl : text;
+      final response = await http.get(Uri.parse(url));
+
+      // Handle the response here
+      if (response.statusCode == 200) {
+        final jsonData = json.decode(response.body);
+        log(jsonEncode(jsonData));
+      } else {
+        log('Request failed with status: ${response.statusCode}');
+      }
+    } catch (e) {
+      log('Error sending request: $e');
+    }
+  }
+}