+++ lib/instabug_flutter.dart
@@ -17,3 +17,5 @@ export 'src/modules/session_replay.dart';
 export 'src/modules/surveys.dart';
 // Utils
 export 'src/utils/instabug_navigator_observer.dart';
+export 'src/utils/screen_loading/instabug_capture_screen_loading.dart';
+export 'src/utils/screen_loading/route_matcher.dart';