+++ lib/src/modules/instabug.dart
@@ -11,11 +11,14 @@ import 'dart:typed_data';
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
+// to maintain supported versions prior to Flutter 3.3
+// ignore: unused_import
 import 'package:flutter/services.dart';
 import 'package:instabug_flutter/instabug_flutter.dart';
 import 'package:instabug_flutter/src/generated/instabug.api.g.dart';
 import 'package:instabug_flutter/src/utils/enum_converter.dart';
 import 'package:instabug_flutter/src/utils/ibg_build_info.dart';
+import 'package:instabug_flutter/src/utils/instabug_logger.dart';
 import 'package:meta/meta.dart';
 
 enum InvocationEvent {
@@ -133,6 +136,8 @@ enum ReproStepsMode { enabled, disabled, enabledWithNoScreenshots }
 class Instabug {
   static var _host = InstabugHostApi();
 
+  static const tag = 'Instabug';
+
   /// @nodoc
   @visibleForTesting
   // ignore: use_setters_to_change_properties
@@ -146,6 +151,18 @@ class Instabug {
     BugReporting.$setup();
     Replies.$setup();
     Surveys.$setup();
+  }
+
+  /// @nodoc
+  @internal
+  static Future<bool> isEnabled() async {
+    return _host.isEnabled();
+  }
+
+  /// @nodoc
+  @internal
+  static Future<bool> isBuilt() async {
+    return _host.isBuilt();
   }
 
   /// Enables or disables Instabug functionality.
@@ -166,6 +183,7 @@ class Instabug {
     LogLevel debugLogsLevel = LogLevel.error,
   }) async {
     $setup();
+    InstabugLogger.I.logLevel = debugLogsLevel;
     return _host.init(
       token,
       invocationEvents.mapToString(),